+++ app/test/e2e/task-get-all.spec.js
@@ -1,17 +1,13 @@
-/* eslint-disable no-unused-vars,no-undef */
 const nock = require('nock');
 const chai = require('chai');
 const Task = require('models/task.model');
-const { createTask, deserializeTask } = require('./utils');
+const { intersection } = require('lodash');
+const { createTask, deserializeTask, validateTask } = require('./utils');
 const { getTestServer } = require('./test-server');
 
-const should = chai.should();
+chai.should();
 
 let requester;
-
-let fakeTask1;
-let fakeTask2;
-let fakeTask3;
 
 nock.disableNetConnect();
 nock.enableNetConnect(process.env.HOST_IP);
@@ -38,9 +34,9 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks should return 200 with the existing tasks', async () => {
-        fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
-        fakeTask2 = await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
-        fakeTask3 = await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+        const fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        const fakeTask2 = await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        const fakeTask3 = await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
 
         const response = await requester
             .get(`/api/v1/doc-importer/task`)
@@ -50,36 +46,70 @@ describe('Task get all tests', () => {
         response.body.should.have.property('data').and.be.an('array').and.have.length(3);
 
         const responseTasks = deserializeTask(response);
-        const task1 = responseTasks[0];
-        const task2 = responseTasks[1];
-        const task3 = responseTasks[2];
 
-        task1.should.have.property('datasetId').and.equal(fakeTask1.datasetId);
-        task1.should.have.property('logs').and.be.an('array').and.have.lengthOf(0);
-        task1.should.have.property('reads').and.equal(0);
-        task1.should.have.property('writes').and.equal(0);
-        task1.should.have.property('message').and.be.an('object');
-        task1.should.have.property('status').and.equal(fakeTask1.status);
-        task1.should.have.property('type').and.equal(fakeTask1.type);
+        validateTask(responseTasks[0], fakeTask1);
+        validateTask(responseTasks[1], fakeTask2);
+        validateTask(responseTasks[2], fakeTask3);
+    });
 
-        task2.should.have.property('datasetId').and.equal(fakeTask2.datasetId);
-        task2.should.have.property('logs').and.be.an('array').and.have.lengthOf(0);
-        task2.should.have.property('reads').and.equal(0);
-        task2.should.have.property('writes').and.equal(0);
-        task2.should.have.property('message').and.be.an('object');
-        task2.should.have.property('status').and.equal(fakeTask2.status);
-        task2.should.have.property('type').and.equal(fakeTask2.type);
+    it('Get a list of existent tasks should return 200 with the existing tasks limited to pagination criteria, and include the pagination metadata', async () => {
+        for (let i = 1; i <= 20; i += 1) {
+            await new Task(createTask('SAVED', 'TASK_CREATE')).save();
+        }
 
-        task3.should.have.property('datasetId').and.equal(fakeTask3.datasetId);
-        task3.should.have.property('logs').and.be.an('array').and.have.lengthOf(0);
-        task3.should.have.property('reads').and.equal(0);
-        task3.should.have.property('writes').and.equal(0);
-        task3.should.have.property('message').and.be.an('object');
-        task3.should.have.property('status').and.equal(fakeTask3.status);
-        task3.should.have.property('type').and.equal(fakeTask3.type);
+        const responseOne = await requester
+            .get(`/api/v1/doc-importer/task`)
+            .query({ page: { number: '1' } })
+            .send();
+
+        responseOne.status.should.equal(200);
+        responseOne.body.should.have.property('data').and.be.an('array').and.have.length(10);
+
+        responseOne.body.should.have.property('links').and.be.an('object');
+        responseOne.body.links.should.have.property('self').and.be.an('string');
+        responseOne.body.links.should.have.property('first').and.be.an('string');
+        responseOne.body.links.should.have.property('last').and.be.an('string');
+        responseOne.body.links.should.have.property('prev').and.be.an('string');
+        responseOne.body.links.should.have.property('next').and.be.an('string');
+
+        responseOne.body.should.have.property('meta').and.be.an('object');
+        responseOne.body.meta.should.have.property('total-pages').and.be.a('number');
+        responseOne.body.meta.should.have.property('total-items').and.be.a('number');
+        responseOne.body.meta.should.have.property('size').and.be.a('number');
+
+        const responseTwo = await requester
+            .get(`/api/v1/doc-importer/task`)
+            .query({ page: { number: '2' } })
+            .send();
+
+        responseTwo.status.should.equal(200);
+        responseTwo.body.should.have.property('data').and.be.an('array').and.have.length(10);
+
+        // Ensure pages contain different tasks
+        intersection(responseOne.body.data.map(task => task.id), responseTwo.body.data.map(task => task.id)).should.eql([]);
+    });
+
+    it('Get a list of existent tasks with a very large page size should return a 400 error code', async () => {
+        for (let i = 1; i <= 20; i += 1) {
+            await new Task(createTask('SAVED', 'TASK_CREATE')).save();
+        }
+
+        const response = await requester
+            .get(`/api/v1/doc-importer/task`)
+            .query({ page: { size: '101' } })
+            .send();
+
+        response.status.should.equal(400);
+        response.body.should.have.property('errors').and.be.an('array').and.have.length(1);
+        response.body.errors[0].should.have.property('status').and.equal(400);
+        response.body.errors[0].should.have.property('detail').and.equal('Invalid page size');
     });
 
     it('Get a list of existent tasks filtered by status should return 200 with the filtered task list', async () => {
+        await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        const fakeTask2 = await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        const fakeTask3 = await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?status=SAVED`)
             .send();
@@ -88,27 +118,16 @@ describe('Task get all tests', () => {
         response.body.should.have.property('data').and.be.an('array').and.have.length(2);
 
         const responseTasks = deserializeTask(response);
-        const task1 = responseTasks[0];
-        const task2 = responseTasks[1];
 
-        task1.should.have.property('datasetId').and.equal(fakeTask2.datasetId);
-        task1.should.have.property('logs').and.be.an('array').and.have.lengthOf(0);
-        task1.should.have.property('reads').and.equal(0);
-        task1.should.have.property('writes').and.equal(0);
-        task1.should.have.property('message').and.be.an('object');
-        task1.should.have.property('status').and.equal(fakeTask2.status);
-        task1.should.have.property('type').and.equal(fakeTask2.type);
-
-        task2.should.have.property('datasetId').and.equal(fakeTask3.datasetId);
-        task2.should.have.property('logs').and.be.an('array').and.have.lengthOf(0);
-        task2.should.have.property('reads').and.equal(0);
-        task2.should.have.property('writes').and.equal(0);
-        task2.should.have.property('message').and.be.an('object');
-        task2.should.have.property('status').and.equal(fakeTask3.status);
-        task2.should.have.property('type').and.equal(fakeTask3.type);
+        validateTask(responseTasks[0], fakeTask2);
+        validateTask(responseTasks[1], fakeTask3);
     });
 
     it('Get a list of existent tasks filtered by type should return 200 with the filtered task list', async () => {
+        await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        const fakeTask3 = await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?type=TASK_OVERWRITE`)
             .send();
@@ -129,6 +148,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by createdAt should return 200 with the filtered task list', async () => {
+        const fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?createdAt=2019-02-01`)
             .send();
@@ -149,6 +172,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by before createdAt should return 200 with the filtered task list', async () => {
+        await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        const fakeTask2 = await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?createdBefore=2019-01-02`)
             .send();
@@ -169,6 +196,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by after createdAt should return 200 with the filtered task list', async () => {
+        const fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        const fakeTask3 = await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?createdAfter=2019-01-02`)
             .send();
@@ -198,6 +229,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by before and after createdAt should return 200 with the filtered task list', async () => {
+        const fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?createdAfter=2019-01-02&createdBefore=2019-02-02`)
             .send();
@@ -218,6 +253,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by updatedAt should return 200 with the filtered task list', async () => {
+        const fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?updatedAt=2019-02-01`)
             .send();
@@ -238,6 +277,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by before updatedAt should return 200 with the filtered task list', async () => {
+        await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        const fakeTask2 = await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?updatedBefore=2019-01-02`)
             .send();
@@ -258,6 +301,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by after updatedAt should return 200 with the filtered task list', async () => {
+        const fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        const fakeTask3 = await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?updatedAfter=2019-01-02`)
             .send();
@@ -287,6 +334,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by before and after updatedAt should return 200 with the filtered task list', async () => {
+        const fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?updatedAfter=2019-01-02&updatedBefore=2019-02-02`)
             .send();
@@ -307,6 +358,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by multiple filters should return 200 with the filtered task list', async () => {
+        const fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?updatedBefore=2019-02-02&status=ERROR`)
             .send();
@@ -327,6 +382,10 @@ describe('Task get all tests', () => {
     });
 
     it('Get a list of existent tasks filtered by dataset id should return 200 with the filtered task list', async () => {
+        const fakeTask1 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+
         const response = await requester
             .get(`/api/v1/doc-importer/task?datasetId=${fakeTask1.datasetId}`)
             .send();
@@ -348,6 +407,11 @@ describe('Task get all tests', () => {
 
 
     it('Get a list of existent tasks should return 200 with the existing tasks, and include details loaded from elasticsearch', async () => {
+        await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_CREATE', new Date('2019-01-01'))).save();
+        await new Task(createTask('SAVED', 'TASK_OVERWRITE', new Date('2019-03-01'))).save();
+        const fakeTask4 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
+
         const elasticTaskResponseObject = {
             completed: true,
             task: {
@@ -397,8 +461,6 @@ describe('Task get all tests', () => {
             }
         };
 
-        const fakeTask4 = await new Task(createTask('ERROR', 'TASK_CREATE', new Date('2019-02-01'))).save();
-
         fakeTask4.logs = [
             {
                 id: '782177aa-fb56-4cf9-bf2c-a9d2cc435a13',
@@ -436,15 +498,13 @@ describe('Task get all tests', () => {
         log.should.have.property('elasticTaskStatus').and.be.an('object').and.deep.equal(elasticTaskResponseObject);
     });
 
-    afterEach(() => {
+    afterEach(async () => {
         if (!nock.isDone()) {
             const pendingMocks = nock.pendingMocks();
             nock.cleanAll();
             throw new Error(`Not all nock interceptors were used: ${pendingMocks}`);
         }
-    });
 
-    after(async () => {
         await Task.deleteMany({}).exec();
     });
 });